fix: require cycle warnings
