+++ dev-client/src/screens/ColorAnalysisScreen/ColorCropReferenceScreen.tsx
@@ -17,7 +17,7 @@
 
 import {useCallback} from 'react';
 import {useTranslation} from 'react-i18next';
-import {useColorAnalysisContext} from 'terraso-mobile-client/screens/ColorAnalysisScreen/ColorAnalysisScreen';
+import {useColorAnalysisContext} from 'terraso-mobile-client/screens/ColorAnalysisScreen/context/colorAnalysisContext';
 import {useColorAnalysisNavigation} from 'terraso-mobile-client/screens/ColorAnalysisScreen/navigation/navigation';
 import {
   ColorCropScreen,
